+++ src/osu_2019/skill_kind.rs
@@ -5,12 +5,12 @@ const SPEED_ANGLE_BONUS_BEGIN: f32 = 5.0 * std::f32::consts::FRAC_PI_6;
 const PI_OVER_4: f32 = std::f32::consts::FRAC_PI_4;
 const PI_OVER_2: f32 = std::f32::consts::FRAC_PI_2;
 
-const MIN_SPEED_BONUS: f64 = 75.0;
-const MAX_SPEED_BONUS: f64 = 45.0;
+const MIN_SPEED_BONUS: f32 = 75.0;
+const MAX_SPEED_BONUS: f32 = 45.0;
 const SPEED_BALANCING_FACTOR: f32 = 40.0;
 
 const AIM_ANGLE_BONUS_BEGIN: f32 = std::f32::consts::FRAC_PI_3;
-const TIMING_THRESHOLD: f64 = 107.0;
+const TIMING_THRESHOLD: f32 = 107.0;
 
 #[derive(Copy, Clone)]
 pub(crate) enum SkillKind {
@@ -19,7 +19,7 @@ pub(crate) enum SkillKind {
 }
 
 impl SkillKind {
-    pub(crate) fn strain_value_of(self, current: &DifficultyObject<'_>) -> f64 {
+    pub(crate) fn strain_value_of(self, current: &DifficultyObject<'_>) -> f32 {
         match self {
             Self::Aim => {
                 if current.base.is_spinner() {
@@ -37,7 +37,7 @@ impl SkillKind {
                             * (current.jump_dist - scale).max(0.0))
                         .sqrt();
 
-                        result = 1.5 * f64::from(apply_diminishing_exp(angle_bonus.max(0.0)))
+                        result = 1.5 * apply_diminishing_exp(angle_bonus.max(0.0))
                             / (TIMING_THRESHOLD).max(prev_strain_time)
                     }
                 }
@@ -45,9 +45,8 @@ impl SkillKind {
                 let jump_dist_exp = apply_diminishing_exp(current.jump_dist);
                 let travel_dist_exp = apply_diminishing_exp(current.travel_dist);
 
-                let dist_exp = f64::from(
-                    jump_dist_exp + travel_dist_exp + (travel_dist_exp * jump_dist_exp).sqrt(),
-                );
+                let dist_exp =
+                    jump_dist_exp + travel_dist_exp + (travel_dist_exp * jump_dist_exp).sqrt();
 
                 (result + dist_exp / (current.strain_time).max(TIMING_THRESHOLD))
                     .max(dist_exp / current.strain_time)
@@ -63,7 +62,7 @@ impl SkillKind {
                 let mut speed_bonus = 1.0;
 
                 if delta_time < MIN_SPEED_BONUS {
-                    let exp_base = (MIN_SPEED_BONUS - delta_time) as f32 / SPEED_BALANCING_FACTOR;
+                    let exp_base = (MIN_SPEED_BONUS - delta_time) / SPEED_BALANCING_FACTOR;
                     speed_bonus += exp_base * exp_base;
                 }
 
@@ -86,11 +85,10 @@ impl SkillKind {
                     }
                 }
 
-                f64::from(
-                    (1.0 + (speed_bonus - 1.0) * 0.75)
-                        * angle_bonus
-                        * (0.95 + speed_bonus * (dist / SINGLE_SPACING_TRESHOLD).powf(3.5)),
-                ) / current.strain_time
+                (1.0 + (speed_bonus - 1.0) * 0.75)
+                    * angle_bonus
+                    * (0.95 + speed_bonus * (dist / SINGLE_SPACING_TRESHOLD).powf(3.5))
+                    / current.strain_time
             }
         }
     }
